archive/backend/genfit_django/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
daphne==4.1.2
channels==4.2.2
groq==0.24.0
requests==2.32.3

